+++ package.json
@@ -11,11 +11,10 @@
     "url": "https://github.com/trueadm/inferno/issues"
   },
   "devDependencies": {
-    "@types/benchmark": "^1.0.30",
     "@types/chai": "^3.4.34",
-    "@types/core-js": "^0.9.35",
     "@types/mocha": "^2.2.33",
     "@types/node": "^6.0.51",
+    "@types/path-to-regexp": "^1.0.29",
     "@types/sinon": "^1.16.33",
     "babel-cli": "^6.18.0",
     "babel-core": "^6.20.0",
@@ -37,7 +36,7 @@
     "babel-register": "^6.18.0",
     "buble": "^0.14.3",
     "chai": "^3.5.0",
-    "concat-stream-es6": "0.0.1",
+    "concat-stream": "^1.5.2",
     "convert-source-map": "^1.3.0",
     "coveralls": "^2.11.15",
     "cross-env": "^3.1.3",
@@ -130,9 +129,8 @@
     "exclude": [
       "packages",
       "src/DOM/devtools.ts",
-      "src/inferno.d.ts",
-      "src/**/*__tests__*/**",
-      "typings"
+      "src/typings/*.d.ts",
+      "src/**/*__tests__*/**"
     ],
     "extension": [
       ".ts"
