+++ packages/inferno/src/index.js
@@ -2,7 +2,7 @@ import {
 	createVNode
 } from '../../../build/core/shapes';
 import cloneVNode from '../../../build/factories/cloneVNode';
-import { warning, NO_OP, isBrowser, EMPTY_OBJ } from '../../../build/shared';
+import { warning, NO_OP, isBrowser, EMPTY_OBJ } from '../../../build/tools/shared';
 import { render, findDOMNode, createRenderer, enableFindDOMNode } from '../../../build/DOM/rendering';
 import { disableRecycling } from '../../../build/DOM/recycling';
 import { initDevToolsHooks } from '../../../build/DOM/devtools';
