+++ src/DOM/utils.ts
@@ -9,7 +9,7 @@ import {
 	isUndefined,
 	throwError,
 	EMPTY_OBJ
-} from '../shared';
+} from '../tools/shared';
 
 import cloneVNode from '../factories/cloneVNode';
 import { componentToDOMNodeMap, findDOMNodeEnabled } from './rendering';
