+++ src/DOM/wrappers/SelectWrapper.ts
@@ -2,7 +2,7 @@ import {
 	isArray,
 	EMPTY_OBJ,
 	isNullOrUndef
-} from '../../shared';
+} from '../../tools/shared';
 import { wrappers } from './processElement';
 import { isVNode } from '../../core/shapes';
